result 변경
