+++ src/js/components/Result.js
@@ -1,51 +1,8 @@
 import { KEY, INIT_STATE } from "../constants.js";
-import { $, createEl } from "../util.js";
+import { $, createEl, calcResult } from "../util.js";
 
 const Result = ({ useEffect }) => {
   const element = createEl.div();
-
-  const calcResult = (lottoList, winningNumbers) => {
-    const result = {
-      3: {
-        value: 0,
-        reward: 5000,
-      },
-      4: {
-        value: 0,
-        reward: 50000,
-      },
-      5: {
-        value: 0,
-        reward: 1500000,
-      },
-      6: {
-        value: 0,
-        reward: 2000000000,
-      },
-      10: {
-        value: 0,
-        reward: 30000000,
-      },
-    };
-
-    for (let lotto of lottoList) {
-      let matchNumber = 0;
-      const isMatchBonus = lotto.includes(Number(winningNumbers[6]));
-
-      for (let i = 0; i < 6; i++) {
-        if (lotto.includes(Number(winningNumbers[i]))) {
-          matchNumber += 1;
-        }
-      }
-      if (matchNumber === 5 && isMatchBonus) {
-        result[10].value += 1;
-      } else if (matchNumber >= 3) {
-        result[matchNumber].value += 1;
-      }
-    }
-
-    return result;
-  };
 
   useEffect(
     (get, set) => {
@@ -53,7 +10,7 @@ const Result = ({ useEffect }) => {
         return (element.innerHTML = null);
       }
 
-      const result = calcResult(get(KEY.LOTTO_LIST), get(KEY.WINNING_NUMBER));
+      const result = get(KEY.RESULT);
 
       element.innerHTML = `<div class="modal">
   <div class="modal-inner p-10">
@@ -127,7 +84,7 @@ const Result = ({ useEffect }) => {
         set(INIT_STATE);
       });
     },
-    [KEY.WINNING_NUMBER, KEY.SHOWING_RESULT]
+    [KEY.RESULT, KEY.SHOWING_RESULT]
   );
 
   return element;
