+++ src/js/constants.js
@@ -1,7 +1,7 @@
 export const KEY = {
   MONEY: "money",
   LOTTO_LIST: "lottoList",
-  WINNING_NUMBER: "winningNumber",
+  RESULT: "result",
   SHOWING_NUMBER: "showingNumber",
   SHOWING_RESULT: "showingResult",
 };
@@ -11,5 +11,5 @@ export const INIT_STATE = {
   [KEY.LOTTO_LIST]: [],
   [KEY.SHOWING_NUMBER]: false,
   [KEY.SHOWING_RESULT]: false,
-  [KEY.WINNING_NUMBER]: [],
+  [KEY.RESULT]: {},
 };
